Flyttade hooks, Settings
Från att logiken leget i App.js har nu den logiken som tillhör Settings lagts där samt att contexten har tagits bort då det ej behövs.
