+++ src/App.js
@@ -2,7 +2,6 @@ import React, {useState} from 'react';
 import Main from './Components/home/MainContentComponent'
 import SideMenu from './Components/SidoMeny/SideMenu'
 import './App.css';
-import SettingsContext from './Components/SettingsContext'
 import PersonalInfo from './Data/PersonalInformation.json'
 
 function App() {
@@ -12,20 +11,7 @@ function App() {
   const [homeClicked, changeHomeState] = useState(true)
   const [portfolioClicked, changePortfolioState] = useState(false)
   const [settingsClicked, changeSettingsState] = useState(false)
-  const [settingOptionsClicked, changeSettingOptionState] = useState(1)
-
-  const settingOptionsProfilClicked = () =>{
-
-        changeSettingOptionState(1);
-  }
-  const settingOptionsPasswordClicked = () =>{
-
-    changeSettingOptionState(2);
-  }
-  const settingOptionsReferenceClicked = () =>{
-
-    changeSettingOptionState(3);
-  }
+ 
 
   const homeButtonClicked = () =>{
     if(!homeClicked){
@@ -52,7 +38,6 @@ function App() {
 
   const dummyProfile = PersonalInfo;
 
-
   return (
     <div className="App">
       <SideMenu
@@ -60,21 +45,11 @@ function App() {
       changePortfolio={portfolioButtonClicked}
       changeSettings={settingsButtonClicked}/>
 
-      <SettingsContext.Provider value= {{
-        myProfilClicked : settingOptionsProfilClicked,
-        changePassClicked : settingOptionsPasswordClicked,
-        referenceClicked : settingOptionsReferenceClicked,
-        currentStateOfSettings : settingOptionsClicked
-      }}>
-
       <Main Date={date.date1}
       homeScreen={homeClicked}
       portfolioScreen={portfolioClicked}
       settingsScreen={settingsClicked}
       person={dummyProfile}/>
-
-      </SettingsContext.Provider>
-
     </div>
   );
 }
