--- src/Components/SettingsContext.js
@@ -1,10 +1,0 @@
-import React from 'react'
-
-const SettingsContext = React.createContext({
-    myProfilClicked : () =>{},
-    changePassClicked : () => {},
-    referenceClicked : () => {},
-    currentStateOfSettings : 1
-
-})
-export default SettingsContext;
